moved routes to MainRouter.tsx
